requirements.txt
Appium-Python-Client==2.6.0
attrs==21.4.0
certifi==2022.6.15
charset-normalizer==2.0.12
idna==3.3
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
pytest-html==3.1.1
pytest-metadata==2.0.1
requests==2.26.0
selenium==4.3.0
tomli==2.0.1
urllib3==1.26.7
wget==3.2
